% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.lmm.R
\name{mice.impute.2l.lmm}
\alias{mice.impute.2l.lmm}
\title{Imputation by a linear mixed model}
\usage{
mice.impute.2l.lmm(y, ry, x, type, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the 
the subset \code{y[ry]} of elements in \code{y} to which the imputation 
model is fitted. The \code{ry} generally distinguishes the observed 
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for 
\code{y}. Matrix \code{x} may have no missing values.}

\item{type}{Vector of length \code{ncol(x)} identifying random and class
variables.  Random variables are identified by a '2'. The class variable
(only one is allowed) is coded as '-2'. Fixed effects are indicated by 
a '1'.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value 
indicates locations in \code{y} for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length 
\code{sum(wy)}
}
\description{
Imputes missing data from a two-level normal model using \code{lme4::lmer()}.
}
\details{
Draws are performed as follows. Using standard asymptotic results, the 
conditional (posterior) distribution of the parameters is approximated by a normal
distribution with mean and variance equal to the estimates of fixed and random effects
and of their variance--covariance, respectively, given restricted maximum likelihood 
(REML) estimates of the variance components.
}
\references{
Moreno-Betancur M., Chavance M. (2016) Sensitivity analysis of incomplete 
longitudinal data departing from the missing at random assumption: Methodology 
and application in a clinical trial with drop-outs.
\emph{Statistical methods in medical research}, 25 (4), 1471-1489
[Epub ahead of print, May 22 2013]
}
\seealso{
Other univariate \code{2l} functions: \code{\link{mice.impute.2l.lmer}},
  \code{\link{mice.impute.2l.norm}},
  \code{\link{mice.impute.2l.pan}}
}
\author{
Margarita Moreno-Betancur (2013)
}
\keyword{datagen}
