% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.logregSensALL.R
\name{mice.impute.logregSensALL}
\alias{mice.impute.logregSensALL}
\alias{mice.impute.logregSens}
\title{For NARFCS procedure only: Not-at-Random Imputation by logistic regression.}
\usage{
mice.impute.logregSensALL(y, ry, x, wy = NULL, xSens, parmS, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Vector of missing data pattern (\code{FALSE}=missing,
\code{TRUE}=observed). This function fits the imputation model to all the data 
(observed and imputed).}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for 
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value 
indicates locations in \code{y} for which imputations are created.}

\item{xSens}{Matrix of covariates for unidentified part of the model.}

\item{parmS}{Vector of sensitivity parameters for the unidentified part of the model.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length 
\code{sum(wy)}
}
\description{
Imputes univariate missing data using using Bayesian logistic regression in 
exactly the same way as in \code{mice.impute.logreg} except the linear predictor
of the logistic model is modified before drawing the binary imputation by adding 
a quantity as specified for the unidentified part of the model through 
\code{predictorSens} or \code{formSens}, and corresponding sensitivity parameter 
values through \code{parmSens}. This results in "not-at-random" imputations.
}
\details{
As for \code{logreg}, perfect prediction is handled by the data augmentation method.
}
\references{
Moreno-Betancur M, Leacy FP, Tompsett D, White I. "mice: The NARFCS procedure for sensitivity analyses" 
(available at: \url{https://rawgit.com/moreno-betancur/NARFCS/master/README.html})

Leacy FP. Multiple imputation under missing not at random assumptions via fully conditional 
specification 2016 (PhD thesis).

Tompsett D, Leacy FP, Moreno-Betancur M, White I. On the use of the not at random fully conditional
specification procedure (NARFCS) in practice (submitted).
}
\seealso{
\code{link{mice.impute.logreg}}

Other univariate imputation functions: \code{\link{mice.impute.cart}},
  \code{\link{mice.impute.lda}},
  \code{\link{mice.impute.logreg.boot}},
  \code{\link{mice.impute.logregSens}},
  \code{\link{mice.impute.logreg}},
  \code{\link{mice.impute.mean}},
  \code{\link{mice.impute.midastouch}},
  \code{\link{mice.impute.norm.boot}},
  \code{\link{mice.impute.norm.nob}},
  \code{\link{mice.impute.norm.predict}},
  \code{\link{mice.impute.normSensALL}},
  \code{\link{mice.impute.normSens}},
  \code{\link{mice.impute.norm}},
  \code{\link{mice.impute.pmm}},
  \code{\link{mice.impute.polr}},
  \code{\link{mice.impute.polyreg}},
  \code{\link{mice.impute.quadratic}},
  \code{\link{mice.impute.rf}},
  \code{\link{mice.impute.ri}}
}
\author{
M. Moreno-Betancur, F. Leacy, D. Tompsett, I. White, 2017
}
\keyword{datagen}
